+++ srcs/ports.c
@@ -6,14 +6,14 @@
 /*   By: cempassi <cempassi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/25 16:21:28 by cempassi          #+#    #+#             */
-/*   Updated: 2021/06/25 16:23:03 by cempassi         ###   ########.fr       */
+/*   Updated: 2021/06/25 16:37:46 by cempassi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_nmap.h"
 #include <stdio.h>
 
-int parse_ports(char *ports)
+uint8_t parse_ports(char *ports)
 {
     printf("Ports string: %s", ports);
     return (EXIT_SUCCESS);
